Configurable kiuwan url in jenkins plugin - Upgrade Jenkins target to 1.579, whole code refactored with retrocompatible configuration support, changed kiuwan client to swagger autogenerated kiuwan client, updated Kiuwan icon
