+++ src/main/java/com/kiuwan/plugins/kiuwanJenkinsPlugin/KiuwanBuildSummaryAction.java
@@ -6,54 +6,38 @@ import org.kohsuke.stapler.DataBoundConstructor;
 import org.kohsuke.stapler.export.Exported;
 import org.kohsuke.stapler.export.ExportedBean;
 
-@ExportedBean(defaultVisibility=2)
+@ExportedBean(defaultVisibility = 2)
 public class KiuwanBuildSummaryAction implements Action {
 
 	private String url;
-	
 	private String icon;
-	
-    @DataBoundConstructor
-    public KiuwanBuildSummaryAction(String url) {
-    	this.url = url;
-    	this.icon = "/plugin/kiuwanJenkinsPlugin/logo.png";
-    }
-    
-    /**
-	 * @return the icon
-	 */
-    @Exported
-    public String getIcon() {
+
+	@DataBoundConstructor
+	public KiuwanBuildSummaryAction(String url) {
+		this.url = url;
+		this.icon = "/plugin/kiuwanJenkinsPlugin/logo.png";
+	}
+
+	@Exported
+	public String getIcon() {
 		return icon;
 	}
-    
-	/**
-	 * @return the url
-	 */
+
 	@Exported
 	public String getUrl() {
 		return url;
 	}
-	
-	/**
-	 * @see hudson.model.Action#getIconFileName()
-	 */
+
 	public String getIconFileName() {
 		return "";
 	}
 
-	/**
-	 * @see hudson.model.Action#getDisplayName()
-	 */
 	public String getDisplayName() {
 		return "";
 	}
 
-	/**
-	 * @see hudson.model.Action#getUrlName()
-	 */
 	public String getUrlName() {
 		return "";
 	}
-    
+
 }
