+++ src/main/java/com/kiuwan/plugins/kiuwanJenkinsPlugin/model/ChangeRequestStatusType.java
@@ -1,0 +1,6 @@
+package com.kiuwan.plugins.kiuwanJenkinsPlugin.model;
+
+public enum ChangeRequestStatusType {
+	RESOLVED,
+	INPROGRESS;
+}
