+++ src/main/java/com/kiuwan/plugins/kiuwanJenkinsPlugin/model/DeliveryType.java
@@ -1,0 +1,6 @@
+package com.kiuwan.plugins.kiuwanJenkinsPlugin.model;
+
+public enum DeliveryType {
+	COMPLETE_DELIVERY,
+	PARTIAL_DELIVERY;
+}
