+++ src/main/java/com/kiuwan/plugins/kiuwanJenkinsPlugin/model/Measure.java
@@ -1,0 +1,7 @@
+package com.kiuwan.plugins.kiuwanJenkinsPlugin.model;
+
+public enum Measure {
+	QUALITY_INDICATOR, 
+	RISK_INDEX, 
+	EFFORT_TO_TARGET;
+}
