+++ src/main/java/com/kiuwan/plugins/kiuwanJenkinsPlugin/model/Mode.java
@@ -1,0 +1,7 @@
+package com.kiuwan.plugins.kiuwanJenkinsPlugin.model;
+
+public enum Mode {
+	STANDARD_MODE,
+	DELIVERY_MODE,
+	EXPERT_MODE;
+}
