+++ src/main/resources/com/kiuwan/plugins/kiuwanJenkinsPlugin/KiuwanBuildSummaryAction/summary.jelly
@@ -1,5 +1,6 @@
+<?jelly escape-by-default='true'?>
 <j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form">
-  <t:summary icon="${rootURL}${it.icon}">
+  <t:summary icon="${it.icon}">
     <a href="${it.url}" target="_blank">View results in kiuwan</a>
   </t:summary>
 </j:jelly>
