+++ src/main/resources/index.jelly
@@ -3,6 +3,7 @@
 
   Since we don't really have anything dynamic here, let's just use static HTML. 
 -->
+<?jelly escape-by-default='true'?>
 <div>
     This plugin integrates <a href="http://www.kiuwan.com" target="_blank">Kiuwan</a> -the cloud static analysis service- to Jenkins.
 </div>
