+++ src/main/webapp/help/help-commandArgs.html
@@ -1,3 +1,7 @@
 <div>
-Arguments to pass to Kiuwan Local Analyzer. <strong>Do not forget to include the double quotes in the beginning and ending of the argument values.</strong></br>For example, to indicate the application name, type the next values:</br><p><pre>	-n "Application name"</pre></p><br /> For more details please refer to <a href="https://www.kiuwan.com/docs/display/K5/Code+analysis+using+the+downloaded+agent">Kiuwan Local Analyzer documentation</a>. 
+Arguments to pass to Kiuwan Local Analyzer.
+<b>Do not forget to include the double quotes in the beginning and ending of the argument values.</b><br>
+For example, to indicate the application name, type the next values:<br>
+<pre>	-n "Application name"</pre>
+<br> For more details please refer to <a href="https://www.kiuwan.com/docs/display/K5/Code+analysis+using+the+downloaded+agent">Kiuwan Local Analyzer documentation</a>.
 </div>
