+++ src/main/webapp/help/help-configureKiuwanURL.html
@@ -1,0 +1,3 @@
+<div>
+	Check this box when under a Kiuwan On Premise environment.
+</div>
