+++ src/main/webapp/help/help-extraParameters.html
@@ -1,3 +1,3 @@
 <div>
-Extra parameters to pass to Kiuwan Local Analyzer, separated by spaces. Allowed format is as follows: &lt;key1&gt;="&lt;value1&gt;" &lt;key2&gt;="&lt;value2&gt;" ... &lt;keyN&gt;="&lt;valueN&gt;". <strong>Do not forget to include the double quotes in the beginning and ending of the parameter value, and to separate each parameter with spaces.</strong>
+Extra parameters to pass to Kiuwan Local Analyzer, separated by spaces. Allowed format is as follows: &lt;key1&gt;="&lt;value1&gt;" &lt;key2&gt;="&lt;value2&gt;" ... &lt;keyN&gt;="&lt;valueN&gt;". <b>Do not forget to include the double quotes in the beginning and ending of the parameter value, and to separate each parameter with spaces.</b>
 </div>
