+++ src/main/webapp/help/help-includes.html
@@ -23,5 +23,4 @@ Rules are:
 <li>If includes provided with no explicit excludes, input file will be accepted only when matches at least one include pattern.</li>
 <li>If both explicit includes and excludes, excludes are checked before includes; a file will be accepted if does not match any exclude pattern but matches at least one include pattern.</li>
 </ol>
-</p>
 </div>
