+++ src/main/webapp/help/help-kiuwanURL.html
@@ -1,0 +1,3 @@
+<div>
+	The Kiuwan URL for your Kiuwan On Premise installation e.g. https://kiuwan.mydomain.com:8888/saas
+</div>
